backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
aiounittest==1.5.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fake-useragent==2.2.0
frozenlist==1.6.2
h11==0.16.0
howlongtobeatpy==1.0.18
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
mysqlclient==2.2.7
pillow==11.2.1
propcache==0.3.1
PyJWT==2.9.0
python-decouple==3.8
redis==6.2.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
gunicorn
